+++ packages/core/src/view/style/edge/EntityRelation.ts
@@ -25,15 +25,17 @@ import type { EdgeStyleFunction } from '../../../types';
 import { EntityRelationConnectorConfig } from '../config';
 
 /**
- * Implements an entity relation style for edges (as used in database
- * schema diagrams). At the time the function is called, the result
- * array contains a placeholder (null) for the first absolute point,
+ * Implements an entity relation style for edges (as used in database schema diagrams).
+ *
+ * At the time the function is called, the result array contains a placeholder (`null`) for the first absolute point,
  * that is, the point where the edge and source terminal are connected.
- * The implementation of the style then adds all intermediate waypoints
- * except for the last point, that is, the connection point between the
- * edge and the target terminal. The first ant the last point in the
- * result array are then replaced with Point that take into account
- * the terminal's perimeter and next point on the edge.
+ *
+ * The implementation of the style then adds all intermediate waypoints except for the last point,
+ * that is, the connection point between the edge and the target terminal.
+ *
+ * The first and the last point in the result array are then replaced with Point that take into account the terminal's perimeter and next point on the edge.
+ *
+ * This EdgeStyle is registered under `entityRelationEdgeStyle` in {@link StyleRegistry} when using {@link Graph} or calling {@link registerDefaultEdgeStyles}.
  *
  * @param state {@link CellState} that represents the edge to be updated.
  * @param source {@link CellState} that represents the source terminal.